debt: Moved settings bits into their own duck
