+++ src/drive/mobile/ducks/settings/components/Unlink.jsx
@@ -2,11 +2,9 @@ import React, { Component } from 'react'
 import { withRouter } from 'react-router'
 import { connect } from 'react-redux'
 import { translate } from 'cozy-ui/react/I18n'
-import SettingCategory, {
-  ELEMENT_BUTTON
-} from '../../components/SettingCategory'
+import SettingCategory, { ELEMENT_BUTTON } from './SettingCategory'
 import Modal from 'cozy-ui/react/Modal'
-import { unlink, getClientSettings } from '../../ducks/authorization'
+import { unlink, getClientSettings } from '../../authorization'
 
 class Unlink extends Component {
   state = {
