+++ src/drive/mobile/ducks/settings/styles.styl
@@ -1,4 +1,5 @@
-@require 'cozy-ui'
+// @require 'cozy-ui'
+@require 'components/forms.styl'
 
 .settings
     margin 0 1rem
@@ -32,3 +33,20 @@
     flex-grow  1
     text-align right
     margin-top 1.25rem
+
+.feedback-form
+    @extend $form
+
+    .coz-form
+        margin-top 0
+
+.feedback-text
+    box-sizing border-box
+    width 100%
+    min-height 12.5rem
+    border .0625rem solid silver
+    padding 1rem
+    margin-bottom 1rem
+
+.button-block
+    text-align right
