--- src/drive/mobile/styles/feedback.styl
@@ -1,19 +1,0 @@
-@require 'components/forms.styl'
-
-
-.feedback-form
-    @extend $form
-
-    .coz-form
-        margin-top 0
-
-.feedback-text
-    box-sizing border-box
-    width 100%
-    min-height 12.5rem
-    border .0625rem solid silver
-    padding 1rem
-    margin-bottom 1rem
-
-.button-block
-    text-align right
